requirements.txt
streamlit
openai-whisper
requests
pytubefix
dotenv